register-api/requirements.txt
pydantic_settings
cwl_utils==0.37
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
ogc-ap-validator==0.5.0
pyyaml
